+++ casper/protocols/blockchain/view_test.go
@@ -6,17 +6,17 @@ import (
 )
 
 func TestView_UpdateSafeEstimates(t *testing.T) {
-	p0 := getProtocol([]uint64{10, 9, 8, 7, 6})
+	p0 := getProtocol([]float64{10, 9, 8, 7, 6})
 	p0.Execute("M-0-A S-1-A M-1-B S-0-B")
 	children0 := make(map[string][]string)
 	children0["A"] = []string{"B"}
 	test(t, children0, p0)
-	p1 := getProtocol([]uint64{10, 9, 8, 7, 6})
+	p1 := getProtocol([]float64{10, 9, 8, 7, 6})
 	p1.Execute("M-0-A S-1-A S-2-A S-3-A S-4-A M-1-B M-2-C M-3-D M-4-E S-0-B S-0-C S-0-D S-0-E")
 	children1 := make(map[string][]string)
 	children1["A"] = []string{"B", "C", "D", "E"}
 	test(t, children1, p1)
-	p2 := getProtocol([]uint64{10, 9, 8, 7, 6})
+	p2 := getProtocol([]float64{10, 9, 8, 7, 6})
 	p2.Execute("M-0-A S-1-A S-2-A M-1-B M-2-C S-0-B S-0-C M-0-D")
 	children2 := make(map[string][]string)
 	children2["A"] = []string{"B", "C"}
